+++ test/BooleanTest.ts
@@ -6,19 +6,19 @@ import { LowercaseBoolean } from "../src/Primitives";
 describe("LowercaseBoolean", () => {
 
     it("matches true", () => {
-        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("true")) as PatternMatch;
+        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("true"), {}) as PatternMatch;
         assert(isPatternMatch(pm));
         assert(pm.$value);
     });
 
     it("matches false", () => {
-        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("false")) as PatternMatch;
+        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("false"), {}) as PatternMatch;
         assert(isPatternMatch(pm));
         assert(!pm.$value);
     });
 
     it("doesn't match junk", () => {
-        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("xyz")) as PatternMatch;
+        const pm = LowercaseBoolean.matchPrefix(inputStateFromString("xyz"), {}) as PatternMatch;
         assert(!isPatternMatch(pm));
     });