+++ test/FloatTest.ts
@@ -7,7 +7,7 @@ describe("Float", () => {
 
     it("test one digit", () => {
         const is = inputStateFromString("1");
-        const m = Float.matchPrefix(is);
+        const m = Float.matchPrefix(is, {});
         expect(isPatternMatch(m)).to.equal(true);
         const match = m as PatternMatch;
         expect(match.$matched).to.equal("1");
@@ -16,7 +16,7 @@ describe("Float", () => {
 
     it("test multiple digits", () => {
         const is = inputStateFromString("105x");
-        const m = Float.matchPrefix(is);
+        const m = Float.matchPrefix(is, {});
         expect(isPatternMatch(m)).to.equal(true);
         const match = m as PatternMatch;
         expect(match.$matched).to.equal("105");
@@ -25,7 +25,7 @@ describe("Float", () => {
 
     it("test with decimal", () => {
         const is = inputStateFromString("105.25555xxx");
-        const m = Float.matchPrefix(is);
+        const m = Float.matchPrefix(is, {});
         expect(isPatternMatch(m)).to.equal(true);
         const match = m as PatternMatch;
         expect(match.$matched).to.equal("105.25555");
@@ -34,7 +34,7 @@ describe("Float", () => {
 
     it("test signed", () => {
         const is = inputStateFromString("-105.25555xxx");
-        const m = Float.matchPrefix(is);
+        const m = Float.matchPrefix(is, {});
         expect(isPatternMatch(m)).to.equal(true);
         const match = m as PatternMatch;
         expect(match.$matched).to.equal("-105.25555");
@@ -43,7 +43,7 @@ describe("Float", () => {
 
     it("test no leading digit", () => {
         const is = inputStateFromString("-.25555xxx");
-        const m = Float.matchPrefix(is);
+        const m = Float.matchPrefix(is, {});
         expect(isPatternMatch(m)).to.equal(true);
         const match = m as PatternMatch;
         expect(match.$matched).to.equal("-.25555");